Preview extraction; js_only.
